+++ mdb/index.go
@@ -46,7 +46,7 @@ func (a *Access) Index(collection *Collection, description *IndexDescription) er
 		},
 	})
 	if err != nil {
-		// TODO: at this point should the collection be removed?
+		// TODO(mAdkins): at this point should the collection be removed?
 		// Experimentation suggests that double creation of the index is OK.
 		return fmt.Errorf("create index on name: %w", err)
 	}
